Replace file link with episode length on list page (MR 1)
This drops the file / .opus link from the index / list page
and instead displays the episode length. This is analogous to
how the approximate reading time is displayed on the pmOS blog.

The episode length is sourced from a header comment in the episode's
markdown, similar to how the timestamps are treated.

Relates to: #3